# server/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shubharambh CRM Enterprise REST API",
    "version": "1.0.0-rc1",
    "description": "Production-grade OpenAPI 3.1 REST API specification for Shubharambh Green City CRM (GIS Township, OCC Plot Booking, MLM Sales Engine & SRE Telemetry).",
    "contact": {
      "name": "Shubharambh Engineering Team",
      "email": "dev@shubharambh.com",
      "url": "https://shubharambh.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Express API Server"
    },
    {
      "url": "https://crm.shubharambh.com/api/v1",
      "description": "Production Kubernetes HTTPS Gateway"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/plots": {
      "get": {
        "summary": "List All Plots (Redis Cached)",
        "tags": ["Plots"],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "block", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "List of plots with status and pricing"
          }
        }
      }
    },
    "/api/v1/booking": {
      "post": {
        "summary": "Create Plot Booking (Server-Authoritative OCC)",
        "tags": ["Booking"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking committed successfully"
          },
          "409": {
            "description": "Optimistic Concurrency Control conflict (Plot already booked)"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateBookingRequest": {
        "type": "object",
        "required": ["plotId", "customerName", "customerPhone", "bookingAmount", "paymentMode", "utrNumber"],
        "properties": {
          "plotId": { "type": "string", "example": "plot-101-guid" },
          "customerName": { "type": "string", "example": "Rajesh Kumar" },
          "customerPhone": { "type": "string", "example": "9876543210" },
          "bookingAmount": { "type": "number", "example": 50000 },
          "paymentMode": { "type": "string", "example": "UPI" },
          "utrNumber": { "type": "string", "example": "UTR998877665544" }
        }
      }
    },
...
  }
}